Adds lcdblib requirements to environment.
As we start to use these wrappers I think it is important the use-conda works.
To get around some of the slowness with building lcdblib I have gone a head and
added the requirments to all of the environment.yaml files that needed them. We
still cannot make use_conda=True default because of the 80 chr name limit would
require renaming all of the test functions.but I have a few tests (bowtie2 aln
se and hisat2 aln se) working with use_conda. Once lcdblib is at a stable
release in conda we can delete the requierments.